[#8] refactor : 업장 정보 입력 / 수정 기능 class명 수정
